requirements.txt
flask==2.3.2
twilio==8.3.0
openai==0.28
python-dotenv==1.0.0